pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.9",
    "langchain>=1.1.2",
    "langchain-google>=0.1.1",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
